Create a data model replicating the JSON files from Factorio.
The current repos that get the data from JSON have way too much logic to
convert things around and to deal with exceptions and stuff. This commit
is the first step in redistributing responsabilities. A set of
repositories will deal with reading the JSON files only and then some
new system will take care of converting all that to structures used by
the optimizer.